+++ src/FreeElephants/JsonApiToolkit/DTO/AbstractAttributes.php
@@ -2,6 +2,10 @@
 
 namespace FreeElephants\JsonApiToolkit\DTO;
 
-abstract class AbstractAttributes extends BaseKeyValueStructure
+/**
+ * @deprecated
+ * @see \FreeElephants\JsonApi\DTO\AbstractAttributes
+ */
+abstract class AbstractAttributes extends \FreeElephants\JsonApi\DTO\AbstractAttributes
 {
 }