+++ src/FreeElephants/JsonApiToolkit/DTO/AbstractDocument.php
@@ -5,30 +5,10 @@ namespace FreeElephants\JsonApiToolkit\DTO;
 use Psr\Http\Message\MessageInterface;
 
 /**
+ * @deprecated
+ * @see \FreeElephants\JsonApi\DTO\AbstractDocument
  * @property AbstractResourceObject|mixed $data
  */
-abstract class AbstractDocument
+abstract class AbstractDocument extends \FreeElephants\JsonApi\DTO\AbstractDocument
 {
-    final public function __construct(array $data)
-    {
-        $concreteClass = new \ReflectionClass($this);
-        $dataProperty = $concreteClass->getProperty('data');
-        /** @var \ReflectionNamedType $reflectionType */
-        $reflectionType = $dataProperty->getType();
-        $dataClassName = $reflectionType->getName();
-        $this->data = new $dataClassName($data['data']);
-    }
-
-    /**
-     * @param MessageInterface $httpMessage
-     * @return static
-     */
-    public static function fromHttpMessage(MessageInterface $httpMessage): self
-    {
-        $httpMessage->getBody()->rewind();
-        $rawJson = $httpMessage->getBody()->getContents();
-        $decodedJson = json_decode($rawJson, true);
-
-        return new static($decodedJson);
-    }
 }