+++ src/FreeElephants/JsonApiToolkit/DTO/BaseKeyValueStructure.php
@@ -2,21 +2,10 @@
 
 namespace FreeElephants\JsonApiToolkit\DTO;
 
-class BaseKeyValueStructure
+/**
+ * @deprecated
+ * @see \FreeElephants\JsonApi\DTO\BaseKeyValueStructure
+ */
+class BaseKeyValueStructure extends \FreeElephants\JsonApi\DTO\BaseKeyValueStructure
 {
-    public function __construct(array $attributes)
-    {
-        $concreteClass = new \ReflectionClass($this);
-        foreach ($attributes as $name => $value) {
-            $property = $concreteClass->getProperty($name);
-            if ($property->hasType()) {
-                $propertyType = $property->getType();
-                if ($propertyType instanceof \ReflectionNamedType && !$propertyType->isBuiltin()) {
-                    $propertyClassName = $propertyType->getName();
-                    $value = new $propertyClassName($value);
-                }
-            }
-            $this->{$name} = $value;
-        }
-    }
 }